Add title() to API (w. initial Action refactoring)
