+++ api-jvm-impl/src/main/java/ch/vorburger/minecraft/storeys/japi/impl/actions/Action.java
@@ -16,30 +16,24 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.vorburger.minecraft.storeys.model;
+package ch.vorburger.minecraft.storeys.japi.impl.actions;
 
-import ch.vorburger.minecraft.storeys.japi.util.Texts;
-import org.spongepowered.api.text.Text;
-import org.spongepowered.api.util.TextMessageException;
+import java.util.concurrent.CompletionStage;
+import javax.annotation.CheckReturnValue;
 
-public class ActionException extends TextMessageException {
+public interface Action<T> {
 
-    private static final long serialVersionUID = 6261204063265579413L;
+    // TODO throws ActionException or return list of ActionValidationViolation-s?
+    // void validate();
 
-    public ActionException(Text message) {
-        super(message);
+    @CheckReturnValue CompletionStage<T> execute(ActionContext context);
+
+    void setParameter(String param);
+
+    default boolean add(Action<?> action) {
+        return false;
     }
 
-    public ActionException(Text message, Throwable throwable) {
-        super(message, throwable);
-    }
-
-    public ActionException(String message) {
-        this(Texts.inRed(message));
-    }
-
-    public ActionException(String message, Throwable throwable) {
-        this(Texts.inRed(message), throwable);
-    }
-
+    // default <T> T requireNonNull(T obj, String propertyName) {
+    // default void checkArgument(boolean test, String propertyValidationErrorMessage) {
 }
