+++ api-jvm-impl/src/main/java/ch/vorburger/minecraft/storeys/japi/impl/actions/ActionException.java
@@ -16,22 +16,30 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.vorburger.minecraft.storeys.util;
+package ch.vorburger.minecraft.storeys.japi.impl.actions;
 
-import java.util.regex.Pattern;
+import ch.vorburger.minecraft.storeys.japi.util.Texts;
+import org.spongepowered.api.text.Text;
+import org.spongepowered.api.util.TextMessageException;
 
-public final class MoreStrings {
+public class ActionException extends TextMessageException {
 
-    private final static Pattern PATTERN = Pattern.compile("(^[\\r\\n]+|[\\r\\n]+$)");
+    private static final long serialVersionUID = 6261204063265579413L;
 
-    private MoreStrings() {
+    public ActionException(Text message) {
+        super(message);
     }
 
-    public static String normalizeCRLF(String text) {
-        return text.replace("\r\n", "\n");
+    public ActionException(Text message, Throwable throwable) {
+        super(message, throwable);
     }
 
-    public static String trimCRLF(String text) {
-        return normalizeCRLF(PATTERN.matcher(text).replaceAll(""));
+    public ActionException(String message) {
+        this(Texts.inRed(message));
     }
+
+    public ActionException(String message, Throwable throwable) {
+        this(Texts.inRed(message), throwable);
+    }
+
 }
