+++ api-jvm-impl/src/main/java/ch/vorburger/minecraft/storeys/japi/impl/actions/ReadingSpeed.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.vorburger.minecraft.storeys;
+package ch.vorburger.minecraft.storeys.japi.impl.actions;
 
 import org.spongepowered.api.text.Text;
