+++ api-jvm-impl/src/test/java/ch/vorburger/minecraft/storeys/japi/impl/util/MoreStringsTest.java
@@ -16,10 +16,10 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.vorburger.minecraft.storeys.util;
+package ch.vorburger.minecraft.storeys.japi.impl.util;
 
-import static ch.vorburger.minecraft.storeys.util.MoreStrings.normalizeCRLF;
-import static ch.vorburger.minecraft.storeys.util.MoreStrings.trimCRLF;
+import static ch.vorburger.minecraft.storeys.japi.util.MoreStrings.normalizeCRLF;
+import static ch.vorburger.minecraft.storeys.japi.util.MoreStrings.trimCRLF;
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
