+++ example/src/main/java/ch/vorburger/minecraft/storeys/example/ExampleScript.java
@@ -32,14 +32,14 @@ public class ExampleScript implements Script {
     @Override public void init(Events e) {
         e.whenCommand("example", m -> {
             // TODO test that this is correctly asynchronously chained - each line wait for execution...
-            m.cmd("/title @s hello, world");
+            m.title("hello, world");
             m.cmd("/tp 232 63 216 -180 25");
-            m.cmd("/narrate Piggy Hello! I'm Piggy.");
+            m.cmd("/narrate Piggy Hello! I'm Piggy. I have a lot to tell you about... do you want to hear it?");
             m.cmd("/say Message in the Chat");
-            m.cmd("/title @s The End");
+            m.title("The End");
         });
         e.whenCommand("another", m -> {
-            m.cmd("/title @s Namaste. Curry pour tous!");
+            m.title("Namaste. Curry pour tous!");
             if (!m.player().getInventory().contains(ItemTypes.FISHING_ROD)) {
                 m.cmd("/say There may be a fishing rod hidden somewhere… look for it, and then catch a fish!");
             } else {
