+++ storeys/src/main/java/ch/vorburger/minecraft/storeys/model/MessageAction.java
@@ -18,6 +18,9 @@
  */
 package ch.vorburger.minecraft.storeys.model;
 
+import ch.vorburger.minecraft.storeys.japi.impl.actions.ActionContext;
+import ch.vorburger.minecraft.storeys.japi.impl.actions.ActionWaitHelper;
+import ch.vorburger.minecraft.storeys.japi.impl.actions.TextAction;
 import java.util.concurrent.CompletionStage;
 import javax.inject.Inject;
