+++ storeys/src/main/java/ch/vorburger/minecraft/storeys/model/NopAction.java
@@ -18,6 +18,8 @@
  */
 package ch.vorburger.minecraft.storeys.model;
 
+import ch.vorburger.minecraft.storeys.japi.impl.actions.Action;
+import ch.vorburger.minecraft.storeys.japi.impl.actions.ActionContext;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
