+++ storeys/src/main/java/ch/vorburger/minecraft/storeys/model/Story.java
@@ -18,6 +18,7 @@
  */
 package ch.vorburger.minecraft.storeys.model;
 
+import ch.vorburger.minecraft.storeys.japi.impl.actions.Action;
 import java.util.List;
 import java.util.stream.Collectors;
