+++ storeys/src/main/java/ch/vorburger/minecraft/storeys/model/parser/CommandMapping.java
@@ -18,7 +18,8 @@
  */
 package ch.vorburger.minecraft.storeys.model.parser;
 
-import ch.vorburger.minecraft.storeys.model.Action;
+import ch.vorburger.minecraft.storeys.japi.impl.actions.Action;
+import ch.vorburger.minecraft.storeys.japi.impl.actions.TitleAction;
 import ch.vorburger.minecraft.storeys.model.AwaitAction;
 import ch.vorburger.minecraft.storeys.model.CommandAction;
 import ch.vorburger.minecraft.storeys.model.DynamicAction;
@@ -26,7 +27,6 @@ import ch.vorburger.minecraft.storeys.model.LocationAction;
 import ch.vorburger.minecraft.storeys.model.MessageAction;
 import ch.vorburger.minecraft.storeys.model.NarrateAction;
 import ch.vorburger.minecraft.storeys.model.NopAction;
-import ch.vorburger.minecraft.storeys.model.TitleAction;
 import com.google.inject.Inject;
 import java.util.ArrayList;
 import java.util.List;
